Speed slider grows quicker after 100% mark

+++ gebr/gebr-job.h
@@ -170,7 +170,7 @@ gchar *gebr_job_get_running_time(GebrJob *job, const gchar *start_date);
 
 gchar *gebr_job_get_elapsed_time(GebrJob *job);
 
-gint gebr_job_get_exec_speed(GebrJob *job);
+gdouble gebr_job_get_exec_speed(GebrJob *job);
 
 void gebr_job_set_exec_speed(GebrJob *job, gdouble exec_speed);
